pyproject.toml
[tool.poetry]
name = "kondo"
version = "0.1.0"
description = ""
authors = ["Ken Erwin <ken@devopslibrary.com>"]

[tool.poetry.dependencies]
python = "^3.5"
requests = "^2.21.0"
python-dateutil = "^2.8.0"
sh = "^1.12.14"
docutils = "^0.14.0"
PyGithub = "^1.43"
gitpython = "^2.1"

[tool.poetry.dev-dependencies]
pytest = "^4.3.0"
pytest-cov = "^2.6.1"
better-exceptions = "*"
docutils = "0.14"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
